[Fix] Main class를 AmongUs Class로 이동 / in-out file arguments 로 지정
